Changes made in emolpyee

+++ RecruitmentManagementPortal/src/main/java/com/rmportal/service/UserServiceImpl.java
@@ -167,7 +167,6 @@ public class UserServiceImpl implements UserServices {
 		User updatedUser = userRepository.findByUserId(id);
 
 		if (Objects.nonNull(updateRequestModel)) {
-			// System.out.println("user details"+user);
 
 			if (UserUtility.isInvalidValue(updateRequestModel.getFirstName())) {
 				throw new CustomException(HttpStatus.NOT_FOUND.value(), "Mandatory Feilds Cannot be Empty");
@@ -215,6 +214,7 @@ public class UserServiceImpl implements UserServices {
 			} else {
 				updatedUser.setDOB(updateRequestModel.getDateOfBirth());
 			}
+			updatedUser.setEmployee_id(updateRequestModel.getEmployee_id());
 			userRepository.save(updatedUser);
 			// conversionUtility.convertForUpdateResponse(user);
 			// userRepository.save(user);
@@ -270,7 +270,7 @@ public class UserServiceImpl implements UserServices {
 		} else {
 			UserToken tokenObj = userTokenRepository.findByTokenValue(token);
 			if (Objects.isNull(tokenObj)) {
-				throw new CustomException(500, " Invalid token");
+				throw new CustomException(500, "Invalid token or user id");
 			}
 
 			/*
@@ -281,6 +281,9 @@ public class UserServiceImpl implements UserServices {
 			 */
 
 			User user = userRepository.findByUserId(userId);
+			if (Objects.isNull(user)) {
+				throw new CustomException(500, "Invalid token or user id");
+			}
 			user.setActive(true);
 			userTokenRepository.delete(tokenObj);
 			return true;
@@ -334,10 +337,10 @@ public class UserServiceImpl implements UserServices {
 	public boolean changePassword(ChangePasswordModel changePasswordModel) throws CustomException {
 		User user = userRepository.findByEmail(changePasswordModel.getEmail());
 		if (Objects.isNull(user))
-			throw new CustomException(500, " Please check mandatory fields");
+			throw new CustomException(500, "Please check mandatory fields");
 
 		if (!user.isActive())
-			throw new CustomException(202, " Password Cannot be changed please contact to Admin");
+			throw new CustomException(202, "Password Cannot be changed please contact to Admin");
 
 		if (!bCryptPassword.matches(changePasswordModel.getOldPassword(), user.getPassword()))
 			throw new CustomException(500, "Old Password did not match");
